[automation] Add bare minimum PromOps automation
Adds static files and steals some of the automation from the kube-ansible
PromOps deployment. Intention here is mostly to provide documentation on
how to set up the environment. You must use kube-ansible to get the environment
running initially, and then pass in the inventory file from that deployment
to allow this to work.

More documentation and changes coming, so there isn't a lot of information
on how to use this currently. Considered experimental and unsupported.