[enhance] Update metrics sample pod to use ConfigMap
Convert the metrics sample pod to using a ConfigMap for the configuration
which will be useful soon when we want to connect it to the QDR. Right now
it simply updates to the new sa_metrics container, and moves from using
an embedded set of command flags in the container, to externalizing
them in the kubernetes manifests.

Adds the -config flag to the Deployment for metrics, and points at a
ConfigMap that is made available via a volumeMount.