Add multi-stage consumer Dockerfiles
Adds a Dockerfile for the metrics and events consumer. This is a multi
stage build that should keep the overall image relatively small. It
makes use of the ONBUILD instruction set to provide an initial build
image for creating the events and metrics consumers.

Also updates the .travis-ci.yml to provide automatic image building
and deployment of images once they are merged into master.

Updates the documentation to better reflect recent changes building
the applications by using dep as a vendoring library manager for the
Golang applications.